Skip downloading if files exist

+++ test/conftest.py
@@ -263,11 +263,15 @@ def get_cifar100():
     data_server = os.environ.get('DATA_SERVER')
     assert data_server
     fn = 'cifar-100-python.tar.gz'
-    url = os.path.join(data_server, fn)
-    logging.info(f'Downloading {fn}')
-    cmd = f'curl -s {url} | tar -zx --strip-components=1 ' \
-         '-C dataset/CIFAR100/cifar-100-python/'
-    execute_cmd(cmd)
+
+    if len(os.listdir('dataset/CIFAR100/cifar-100-python/')) >= 5:
+        logging.info(f'{fn} already downloaded')
+    else:
+        url = os.path.join(data_server, fn)
+        logging.info(f'Downloading {fn}')
+        cmd = f'curl -s {url} | tar -zx --strip-components=1 ' \
+             '-C dataset/CIFAR100/cifar-100-python/'
+        execute_cmd(cmd)
 
 @pytest.fixture(scope='session')
 def get_imagenet_val():
@@ -275,33 +279,44 @@ def get_imagenet_val():
     assert data_server
     fn = 'ILSVRC2012_img_val.tar'
     url = os.path.join(data_server, fn)
+    dst = 'dataset/ILSVRC2012/ILSVRC2012_img_val/'
+    if len(os.listdir(dst)) >= 50000:
+        logging.info(f'{fn} already downloaded')
+        return
     logging.info(f'Downloading {fn}')
-    cmd = f'curl -s {url} | tar -x -C dataset/ILSVRC2012/ILSVRC2012_img_val/'
+    cmd = f'curl -s {url} | tar -x -C {dst}'
     execute_cmd(cmd)
 
 @pytest.fixture(scope='session')
 def get_coco2017_val():
     data_server = os.environ.get('DATA_SERVER')
     assert data_server
+
     fn = 'val2017.zip'
     url = os.path.join(data_server, fn)
-    logging.info(f'Downloading {fn}')
-    cmd = f'curl -o val2017.zip -s {url}'
-    execute_cmd(cmd)
-    cmd = 'unzip -o val2017.zip -d dataset/COCO2017'
-    execute_cmd(cmd)
-    cmd = 'rm val2017.zip'
-    execute_cmd(cmd)
+    if len(os.listdir('dataset/COCO2017/val2017')) >= 5000:
+        logging.info(f'{fn} realdy downloaded')
+    else:
+        logging.info(f'Downloading {fn}')
+        cmd = f'curl -o val2017.zip -s {url}'
+        execute_cmd(cmd)
+        cmd = 'unzip -o val2017.zip -d dataset/COCO2017'
+        execute_cmd(cmd)
+        cmd = 'rm val2017.zip'
+        execute_cmd(cmd)
 
     fn = 'annotations_trainval2017.zip'
-    url = os.path.join(data_server, fn)
-    logging.info(f'Downloading {fn}')
-    cmd = f'curl -o annotations.zip -s {url}'
-    execute_cmd(cmd)
-    cmd = 'unzip -o annotations.zip -d dataset/COCO2017/'
-    execute_cmd(cmd)
-    cmd = 'rm annotations.zip'
-    execute_cmd(cmd)
+    if len(os.listdir('dataset/COCO2017/annotations')) >= 7:
+        logging.info(f'{fn} realdy downloaded')
+    else:
+        url = os.path.join(data_server, fn)
+        logging.info(f'Downloading {fn}')
+        cmd = f'curl -o annotations.zip -s {url}'
+        execute_cmd(cmd)
+        cmd = 'unzip -o annotations.zip -d dataset/COCO2017/'
+        execute_cmd(cmd)
+        cmd = 'rm annotations.zip'
+        execute_cmd(cmd)
 
 def main():
     logging.basicConfig(level=logging.INFO)
